Slim diagnostics classes, rename GridDiagnostics to GridModelDiagnostics
Renames for consistency with the class publishing it - `<ClassName>Diagnostics` across all three.

Strips the TSDoc. These are internal, deliberately unstable APIs, and the prose was outweighing
the ~30 lines of declarations it described. Each `diagnostics` property keeps a one-line comment
noting what it holds and that it is not a stable API.

Inlines the op `type` unions into their interfaces, dropping the separate `StoreOpType` /
`ViewOpType` / `GridOpType` exports. The two internal call sites needing the union now use indexed
access (`StoreOp['type']`).
