+++ CHANGELOG.md
@@ -121,7 +121,8 @@ configs for zero-copy projection, digest-based record reuse, and streaming loads
   the path it took - including, for a `View`, the specific condition that forced a full row
   regeneration rather than an incremental update. Read across the three stages, these localize the
   cost of a data change to the stage responsible for it. Note these objects are intentionally NOT a
-  stable API - see `StoreDiagnostics`, `ViewDiagnostics`, and `GridDiagnostics`.
+  stable API - their shape, and especially the set of `type` values, tracks Hoist internals and can
+  change in any release.
 * Removed `Store.patchStats`, superseded by `Store.diagnostics`. The new API reports the last
   operation of each kind rather than cumulative counters - loads, updates, and filter runs differ
   by orders of magnitude in cost, so an average across them was not meaningful - and unlike
