+++ cmp/grid/GridModel.ts
@@ -110,7 +110,7 @@ import {
 import {computed} from 'mobx';
 import {createRef, ReactNode, RefObject} from 'react';
 import {GridAutosizeOptions} from './GridAutosizeOptions';
-import {GridDiagnostics} from './GridDiagnostics';
+import {GridModelDiagnostics} from './GridModelDiagnostics';
 import {GridContextMenuItemLike, GridContextMenuSpec} from './GridContextMenu';
 import {GridSorter, GridSorterLike} from './GridSorter';
 import {initPersist} from './impl/InitPersist';
@@ -623,11 +623,8 @@ export class GridModel extends HoistModel {
     /** Tracks execution of autosize operations. */
     @managed autosizeTask = TaskObserver.trackAll();
 
-    /**
-     * Detail on the last transaction this grid synced to ag-Grid, for performance debugging and
-     * developer tooling. Not a stable API - see {@link GridDiagnostics}.
-     */
-    readonly diagnostics = new GridDiagnostics();
+    /** Detail on the last transaction synced to ag-Grid. Not a stable API. */
+    readonly diagnostics = new GridModelDiagnostics();
 
     constructor(config: GridConfig) {
         super();
